+++ desktop/src/renderer/src/components/index.ts
@@ -1,3 +1,4 @@
 export * from "./VPNPopup";
 export * from "./AuthButton";
 export * from "./SideNav";
+export * from "./Toast";
